refactor: Move remaining logic in one big move
