+++ src-tauri/src/northstar/mod.rs
@@ -1,9 +1,8 @@
 //! This module deals with handling things around Northstar such as
 //! - getting version number
 
-use crate::constants::CORE_MODS;
+use crate::{check_origin_running, constants::CORE_MODS, get_host_os, GameInstall, InstallType};
 use anyhow::anyhow;
-use app::{check_origin_running, get_host_os, GameInstall, InstallType};
 
 /// Check version number of a mod
 pub fn check_mod_version_number(path_to_mod_folder: &str) -> Result<String, anyhow::Error> {
