+++ src-tauri/src/repair_and_verify/mod.rs
@@ -1,7 +1,7 @@
 use crate::mod_management::{get_enabled_mods, rebuild_enabled_mods_json, set_mod_enabled_status};
-use anyhow::anyhow;
 /// Contains various functions to repair common issues and verifying installation
-use app::{constants::CORE_MODS, GameInstall};
+use crate::{constants::CORE_MODS, GameInstall};
+use anyhow::anyhow;
 
 /// Verifies Titanfall2 game files
 #[tauri::command]
